+++ src/music/mod.rs
@@ -1,1 +1,2 @@
+pub mod chord_name;
 pub mod scale;